+++ packages/plugin-proxy/test/fixtures/docs-example/routes/proxy.js
@@ -6,16 +6,16 @@ module.exports = [
     variants: [
       {
         id: "enabled",
-        handler: "proxy-v4",
-        response: {
+        type: "proxy-v4",
+        options: {
           host: "http://127.0.0.1:3200",
           options: {},
         },
       },
       {
         id: "disabled",
-        handler: "middleware",
-        response: {
+        type: "middleware",
+        options: {
           middleware: (_req, _res, next) => next(),
         },
       },