+++ packages/core/CHANGELOG.md
@@ -12,12 +12,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-### Changed
-- refactor(#334): Reorganize files and folders
-- feat(#334): Deprecate `core.restartServer`, `core.addRouter` and `core.removeRouter` methods. Add alerts when used. `core.server` methods must be used instead.
+### Added
+
 - feat(#334): Expose new `core.mock` API
 - feat(#334): Add new option `mock.routes.delay`. Deprecate `mocks.delay`
 - feat(#334): Add new option `mock.collections.selected`. Deprecate `mocks.selected`
+- feat(#334): Support `type` and `options` properties in variants
+
+### Changed
+- refactor(#334): Reorganize files and folders
+- feat(#334): Deprecate `core.restartServer`, `core.addRouter` and `core.removeRouter` methods. Add alerts when used. `core.server` methods must be used instead.
 - feat(#334): Deprecate `core.addRoutesHandler` method. Add alerts when used. `core.variantHandlers.register` method must be used instead.
 - feat(#334): Deprecate `routesHandlers` option. `variantHandlers.register` option must be used instead.
 - feat(#334): Deprecate `core.mocks.restoreRoutesVariants` method. `core.mock.restoreRouteVariants` method must be used instead.